Added img url inputs
Added two input boxes for the user to enter an image URL for their headshot and an image URL for their alma mater on the sign-up page. Added a comment briefly explaining the displayNewContentAfter2Seconds function in the relationships.html page. Added a few comments in the C# file and the signUp.html page to explain the acronyms "hs" and "am."